Separate donor and nonprofit history
