+++ app/dashboard/organization/history/nonProfitHistory.tsx
@@ -1,0 +1,13 @@
+import PageHeading from "@/components/dashboard/PageHeading";
+
+// TODO get grants applied to and display similar to donor history
+export default function NonProfitHistory() {
+  return (
+    <div>
+      <PageHeading
+        header="Grant History"
+        description="History of previous grants received"
+      />
+    </div>
+  );
+}
